DRY Violation via: Commented out 2 lines of code that set a default form field value on the character form. These are not needed as this is set in the mainwindow.ui as the default value.
A test of the form load showed the correct value present on load from
the ui file. Regression testing in other areas was not completed on this
change yet due to my lack of knowledge of all the uses of this symbol. I
do believe this will cause NO issue however since the default is set
correctly prior to it becoming available to other code.